refactor: Added create game logic to the UI file for both Task and Game;
