+++ index.html
@@ -1,4 +1,4 @@
-<!-- index.html file -->
+<!-- index.html -->
 <!DOCTYPE html>
 <html lang="pt-br">
     <head>
