+++ js/utils/validationUtils.js
@@ -1,3 +1,4 @@
+// js\utils\validationUtils.js
 export function validateNumberInput(input) {
     input.value = input.value.replace(/[^0-9]/g, '');  // Removes non-numeric characters (.,)
 }
