+++ tests/dbInit.test.js
@@ -1,3 +1,4 @@
+// tests\dbInit.test.js
 import Dexie from 'dexie';
 
 let db;
